requirements.txt
colorama
pycryptodome
pywin32